Refactored viewHeaderRowsMenu component
